% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showResult.R
\name{changeModelParameter}
\alias{changeModelParameter}
\title{change the model parameter}
\usage{
changeModelParameter(parameterValue, paraIndex, value)
}
\arguments{
\item{parameterValue}{the parameter value series which corresponded to the parameter name series}

\item{paraIndex}{the name of parameter which will be modified}

\item{value}{the new value of parameter which will be modified}
}
\value{
the new parameter series
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
CHEN　Longzan
}
