% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para.R
\name{check_basinData}
\alias{check_basinData}
\title{the basin data integrity test}
\usage{
check_basinData(basinData)
}
\arguments{
\item{basinData}{the basin data}
}
\value{
if everything is ok , there will be no return information
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
CHEN Longzan
}
