% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dXAJ.R
\name{dXAJ}
\alias{dXAJ}
\title{The day model of XAJ}
\usage{
dXAJ(dayModelParameter, basinInfo, dayData)
}
\arguments{
\item{dayModelParameter}{the model parameter series which corresponded to the parameter name series}

\item{basinInfo}{the basin information}

\item{dayData}{the day model data series}
}
\value{
resultData which including date,precipitation,calculated discharge,measured discharge
}
\description{
\description{} The Day model of XINANJIANG hydrology model
}
\details{
\details{} This function is the entrance of day-model, you may use it directly.
}
\author{
CHEN Longzan
}
