% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hXAJ.r
\name{hXAJ}
\alias{hXAJ}
\title{The flood model of XAJ model}
\usage{
hXAJ(hourModelParameter, basinInfo, floodData)
}
\arguments{
\item{hourModelParameter}{the model parameter}

\item{basinInfo}{the basin information}

\item{floodData}{the hour model data series}
}
\value{
floodResult: the hour model result series
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
CHEN Longzan
}
