% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initHourData.r
\name{initHourData}
\alias{initHourData}
\title{initHourData}
\usage{
initHourData(dayStart, floodData)
}
\arguments{
\item{dayStart}{the date of the hour-model start}

\item{floodData}{the whole flood data}
}
\value{
the new flood data
}
\description{
\description{}  This function provided the initial data which need by hour model. provided the initial data which need by hour model
}
\details{
\details{} This function needs the input including the date of the flood start and the floodData which should be generated at the beginning.
}
\author{
CHEN Longzan
}
