% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showResult.R
\name{NSE}
\alias{NSE}
\title{This function is used for calculating the Nash-Sutcliffe efficiency coefficient}
\usage{
NSE(Qcal, Qmea)
}
\arguments{
\item{Qcal}{the calculated discharge}

\item{Qmea}{the measured discharge}
}
\value{
the Nash-Sutcliffe efficiency coefficient
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
CHEN Longzan
}
