% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dXAJ.R
\name{dXAJ}
\alias{dXAJ}
\title{day model of XAJ}
\usage{
dXAJ(modelParameter, basinInfo, basinData)
}
\arguments{
\item{modelParameter}{the model parameter series which corresponded to the parameter name series}

\item{basinInfo}{the basin information}

\item{basinData}{the day model data series}
}
\value{
resultData which including date,precipitation,calculated discharge,measured discharge
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
CHEN Longzan
}
