% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hXAJ.r
\name{hXAJ}
\alias{hXAJ}
\title{the hour model of XAJ model}
\usage{
hXAJ(modelParameter, basinInfo, basinData)
}
\arguments{
\item{modelParameter}{the model parameter}

\item{basinInfo}{the basin information}

\item{basinData}{the hour model data series}
}
\value{
the hour model result series
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
CHEN Longzan
}
