% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initHourData.r
\name{initHourData}
\alias{initHourData}
\title{provided the initial data which need by hour model}
\usage{
initHourData(dayStart, floodData)
}
\arguments{
\item{dayStart}{the date of the hour-model start}

\item{floodData}{the whole flood data}
}
\value{
the new flood data
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
CHEN Longzan
}
