% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showResult.R
\name{showResult}
\alias{showResult}
\title{this function is used for showing the simulation in chart}
\usage{
showResult(Date, P, Qcal, Qmea)
}
\arguments{
\item{Date}{the date series}

\item{P}{the precipitation series}

\item{Qcal}{the calculated discharge series}

\item{Qmea}{the measured discharge series}
}
\value{
a chart object which will be showed in Plots windows
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
CHEN Longzan
}
